Create TicTacToe3DEnglish.java
This is a version of the TicTacToe game that is meant to represent a 3D grid, but instead uses "layers" (because 3D graphics = hard). It also has GUI (Graphical User Interface) because it makes it a bit easier to play instead of typing the x,y,z in the cartesian coordinate system. The game is pretty basic but I hope you like it!

*Note: The six private boolean methods can be shortened to just four methods (the same way that is used in the TicTacToe4Dabove class). But in my mind it makes it easier to read (when the dimensions are only three)