Create TicTacToe4Dabove.java
This is by far the most extended version of the original Tic-Tac-Toe game I have made. It can be customized in many criterias including:
Number of players, number of dimensions, the length of each dimension, (planning to add color to each player's symbol), the winning length (how many pieces you need to put horizontally, vertically or diagonally in order to win).

The mechanics are straight-forward. If you wonder how I store the characters of an N dimensional board - I dont just make N dimensional char array. That would be stupid and unnecessary because then to do the four interconnected check win methods you would have to do N nested for loops for each char in each dimension. Instead I use a technique called "column-major ordering". It basically appropriately converts the multidimensional array into a one dimensional array which is way easier to work with. To fully grasp the idea of this technique, I will recommend checking out this links:
https://en.wikipedia.org/wiki/Row-_and_column-major_order
(go to paragraph - "Address calculation in general" to see the general formula that I used to convert the coordinates of each char in the N space to a single index and vise versa. There you can see a proof that these are one-to-one functions also called bijective functions. That means that the input of each one of them is processed in a way such that one input to the function yields a unique output. This is needed because if this wasn't the case, then there would be two or more symbols that match to the same index and then the inverse function wouldn't work, therefore ruining the check-win)
https://math.stackexchange.com/questions/2008367/how-to-convert-an-index-into-n-coordinates/2008544#comment10782931_2008544
(This link has an answer that explains in detail the operation of the method and even gives some examples in other programming languages.)